requirements.txt
streamlit
pdfplumber
python-dotenv
scikit-learn
sentence-transformers
google-genai
langgraph
langchain-core
torch